Refactor: Complete import verification and fixes.
This commit addresses the 'Imports and Scripts Verification' task.

Key changes include:

1.  **Standardized Imports**: Ensured all imports across `apps/agents/`, `apps/workflows/`, `core/`, and `tests/` are absolute from the project root.
2.  **`__init__.py` Management**: Verified and created `__init__.py` files as needed in all relevant directories, particularly within `tests/` subdirectories, to ensure proper package recognition.
3.  **Circular Dependency Resolution (`core` <-> `apps`)**:
    *   Moved `BaseAgent` and `AgentCapability` from `apps.agents.base_agent.py` to a new file `core.base_agent_abc.py`.
    *   Updated `core.agent_registry.py` to import `BaseAgent` from its new location in `core`.
    *   Updated all agent implementations in `apps.agents.*` and relevant test files to import `BaseAgent` and `AgentCapability` from `core.base_agent_abc`.
4.  **SQLAlchemy `Base` Correction**:
    *   Modified `core.database.orm_models.py` to use the shared SQLAlchemy `Base` object defined in `core.database.base.py`, removing its local `declarative_base()` definition.
5.  **Minor Import Fixes**:
    *   Added a missing import for `AgentProcessingError` in `apps.agents.decision.scheduling_agent.py`.

These changes improve the overall code hygiene, enforce a clearer architectural separation (especially between `core` and `apps`), and resolve potential runtime issues related to inconsistent imports or circular dependencies. The project's import structure is now verified and robust.
