+++ smart-maintenance-saas/core/database/orm_models.py
@@ -13,10 +13,9 @@ from sqlalchemy import (
     Text,
 )
 from sqlalchemy.dialects.postgresql import ARRAY, JSONB, UUID
-from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import relationship
 
-Base = declarative_base()
+from core.database.base import Base
 
 
 class SensorReadingORM(Base):
