+++ smart-maintenance-saas/tests/integration/core/__init__.py
@@ -1,0 +1,1 @@
+"""Integration tests for core components."""
